Update api definition to match current approach
